Lots-o-Nesting
I know this looks bad... I want to use a slug and just dynamically make the page. It'd way more elegant and flexible. But frankly this is faster and idk if they're even going to USE this thing, so, insane nesting for now and I'll make the change it to a dynamic slug down the road if that's what happens. If anyone is reading this, dont judge me for the insane nesting.